Preparing for Linux support

+++ Simplex/src/Native/OpenGL/OpenGLIndexBuffer.cpp
@@ -12,7 +12,7 @@ OpenGLIndexBuffer::OpenGLIndexBuffer(IndexBufferProps props)
     if (!SXG::TypeIsInteger(props.indexType))
         LOG_ERROR("Index buffer should consist of integral type.");
 
-    m_Count = props.size / SXG::TypeSize(props.indexType);
+    m_Count = (unsigned int) (props.size / SXG::TypeSize(props.indexType));
     m_Type = props.indexType;
 
     glCreateBuffers(1, &m_ContextID);
